+++ graphql_client_codegen/src/codegen/shared.rs
